refactor: move GrantType to common/types/billing.ts
- Move GrantType from db/schema to common/types/billing.ts
- Update imports to use new type location
- Keep implementation details in backend
- Maintain npm-app isolation from db access
