+++ backend/src/websockets/websocket-action.ts
@@ -9,7 +9,7 @@ import { sendMessage } from './server'
 import db from 'common/db'
 import * as schema from 'common/db/schema'
 import { protec } from './middleware'
-import { calculateUsageAndBalance } from 'common/src/billing/balance-calculator'
+import { calculateUsageAndBalance } from '../billing/balance-calculator'
 import { ensureEndsWithNewline } from 'common/src/util/file'
 import { logger, withLoggerContext } from '@/util/logger'
 import { generateCompactId } from 'common/util/string'
